+++ spec/gridstack-engine-spec.ts
@@ -359,27 +359,32 @@ describe('gridstack engine', function() {
     });
     it('should add widgets around locked one', function() {
       let nodes: GridStackNode[] = [
-        {x: 0, y: 1, w: 12, h: 1, locked: true, noMove: true, noResize: true, id: 1},
-        {x: 1, y: 0, w: 2, h: 3, id: 2}
+        {x: 0, y: 1, w: 12, h: 1, locked: true, noMove: true, noResize: true, id: 0},
+        {x: 1, y: 0, w: 2, h: 3, id: 1}
       ];
       // add locked item
       engine.addNode(nodes[0])
-      expect(findNode(engine, 1)).toEqual(jasmine.objectContaining({x: 0, y: 1, w: 12, h: 1, locked: true}));
-      engine.addNode(nodes[1])
+      expect(findNode(engine, 0)).toEqual(jasmine.objectContaining({x: 0, y: 1, w: 12, h: 1, locked: true}));
       // add item that moves past locked one
-      expect(findNode(engine, 1)).toEqual(jasmine.objectContaining({x: 0, y: 1, w: 12, h: 1, locked: true}));
-      expect(findNode(engine, 2)).toEqual(jasmine.objectContaining({x: 1, y: 2, h: 3, id: 2}));
-      // prevents moving locked item
+      engine.addNode(nodes[1])
+      expect(findNode(engine, 0)).toEqual(jasmine.objectContaining({x: 0, y: 1, w: 12, h: 1, locked: true}));
+      expect(findNode(engine, 1)).toEqual(jasmine.objectContaining({x: 1, y: 2, h: 3}));
+      // locked item can still be moved directly (what user does)
+      let node0 = findNode(engine, 0);
+      expect(engine.moveNode(node0, {y:6})).toEqual(true);
+      expect(findNode(engine, 0)).toEqual(jasmine.objectContaining({x: 0, y: 6, h: 1, locked: true}));
+      // but moves regular one past it
       let node1 = findNode(engine, 1);
-      expect(engine.moveNode(node1, {x:6, y:6})).toEqual(false);
-      // but moves regular one (gravity ON)
-      let node2 = findNode(engine, 2);
-      expect(engine.moveNode(node2, {x:6, y:6})).toEqual(true);
-      expect(node2).toEqual(jasmine.objectContaining({x: 6, y: 2, w: 2, h: 3}));
-      // but moves regular one (gravity OFF)
+      expect(engine.moveNode(node1, {x:6, y:6})).toEqual(true);
+      expect(node1).toEqual(jasmine.objectContaining({x: 6, y: 7, w: 2, h: 3}));
+      // but moves regular one before (gravity ON)
+      engine.float = false;
+      expect(engine.moveNode(node1, {x:7, y:3})).toEqual(true);
+      expect(node1).toEqual(jasmine.objectContaining({x: 7, y: 0, w: 2, h: 3}));
+      // but moves regular one before (gravity OFF)
       engine.float = true;
-      expect(engine.moveNode(node2, {x:7, y:6})).toEqual(true);
-      expect(node2).toEqual(jasmine.objectContaining({x: 7, y: 6, w: 2, h: 3}));
+      expect(engine.moveNode(node1, {x:7, y:3})).toEqual(true);
+      expect(node1).toEqual(jasmine.objectContaining({x: 7, y: 3, w: 2, h: 3}));
     });
   });